Add performance monitoring of PWM generation on C2000
Create a simple performance monitor class that uses TIMER
0 as a cycle counter running at SYSCLOCK frequency. Use
this to measure the time taken to execute the
PwmGeneration::Run() method and report this back through
the PwmDriver.

Tests:
 - Run Debug and Release builds and verify that the numbers
   make some sort of sense.
 - Test the duration of a 0.5 second delay using the
   performance counter and verify this equates to the correct
   number of cycles when running at 200MHz.

+++ platform/c2000/inverter/c2000pwmdriver.h
@@ -49,7 +49,7 @@ public:
     static void    AcHeatTimerSetup();
     static void    AcHeat(s32fp ampnom);
     static void    SetChargeCurrent(int16_t dc);
-    static int16_t GetCpuLoad();
+    static int32_t GetCpuLoad();
     static void    ResetCpuLoad();
 };
